requirements.txt
# Pinterest Test Automation Framework
# Requirements file

# Core Testing
pytest>=7.4.0
pytest-html>=4.1.1
pytest-xdist>=3.5.0        # Parallel execution
pytest-timeout>=2.2.0      # Test timeouts
pytest-rerunfailures>=13.0 # Retry flaky tests

# Browser Automation
playwright>=1.40.0

# API Testing
requests>=2.31.0
urllib3>=2.1.0

# Utilities
python-dotenv>=1.0.0       # Environment variables
pydantic>=2.5.0            # Data validation (optional)

# Code Quality
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0

# Reporting
allure-pytest>=2.13.2      # Allure reports (optional)
